Editor client: Fix property type in property grid
