+++ crates/lgn-editor-gui/frontend/tests/components/propertyGrid/PropertyGrid.test.ts
@@ -1,8 +1,9 @@
 import { cleanup } from "@testing-library/svelte";
 import { render } from "@testing-library/svelte";
 
+import type { ResourceProperty } from "@lgn/proto-editor/dist/property_inspector";
+
 import PropertyGrid from "@/components/propertyGrid/PropertyGrid.svelte";
-import type { ResourceProperty } from "@/lib/propertyGrid";
 import { formatProperties } from "@/lib/propertyGrid";
 import currentResource from "@/orchestrators/currentResource";
 import properties from "@/resources/propertiesResponse.json";
@@ -36,6 +37,7 @@ describe("PropertyGrid", () => {
         type: "",
       },
       version: 1,
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
       properties: formatProperties(properties as unknown as ResourceProperty[]),
     });
